% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Example leaflet widget}
\usage{
make_map(title = "This is a test", lat, lng)
}
\arguments{
\item{title}{label for the marker}

\item{lat}{lattitude coordinate}

\item{lng}{longintude coordinate}
}
\description{
Hello world of leaflet htmlwidget wrapped in OpenCPU. Based on example:
\url{http://rstudio.github.io/leaflet/}.
}

